Delete first and last month for each user
Previously deleted last month from dataset and ignored first and last
month when selecting for sufficient txns per month. Problem with this
was that there were months in data with fewer then min txns or spend if
they were first or last month in a user's data. This was problematic
because it biases entropy scores towards zero.

Now, I simply remove first and last month for each user, and then drop
users with fewer than min txns in any month.