Fix analysis data bugs
Require region as demographic info rather than postcode so as to
automatically check for valid postcodes (which are those for whom we can
find a region).

Use category variable for inflows and outflows to get zero inflows in
cases where there are none in the data. Previously, code relied on there
being some inflows.